requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Telegram
python-telegram-bot==20.7

# Yandex Tracker
yandex-tracker-client==2.9

# LLM
requests==2.31.0
openai==1.3.7

# Scheduler
apscheduler==3.10.4

# Logging
python-json-logger==2.0.7

asyncpg==0.29.0
httpx==0.25.2

jinja2>=3.1.0 